# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: IITM Beacon API
  description: >
    Exchange Student Testimonial & Review Gallery. Covers the six feature slices: gallery
    (public browse/filter/search), submission (visitor OTP login, session ping,
    create/edit-testimonial), moderation (admin queue/approve/reject, one at a time, session
    ping), catalogadmin (admin CRUD for topic groups/topics/achievements), adminauth
    (passwordless OTP login), analytics (homepage dashboard). Revealing a testimonial's public
    contact methods (UC-REVEAL-CONTACT) is deliberately not part of this JSON API: it is only
    served to the site's own article page, as the HTML route `POST /gallery/{id}/contact`
    (decision 27). See ../use-cases.md for the scenarios behind each endpoint,
    ../decisions.md for the design decisions behind each schema shape, and ../nfr.md for the
    quality attributes they must meet.
  version: "0.3.0"
servers:
  - url: /api
...
paths:
...
  /submissions/otp/request:
    post:
      tags: [submission]
      summary: Request a visitor login OTP (UC-VISITOR-LOGIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202":
          description: >
            Accepted regardless of whether a testimonial already exists for this email —
            identical response either way, so requesting one never reveals that on its own
            (decision 17, NFR-VISITOR-OTP-BRUTEFORCE).
        "429":
          description: Too many OTP requests in a short window, for this email or this IP.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /submissions/otp/verify:
    post:
      tags: [submission]
      summary: Verify a visitor login OTP and establish a session (UC-VISITOR-LOGIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, minLength: 6, maxLength: 6 }
      responses:
        "200":
          description: >
            OTP correct — session cookie set. `mode` tells the client whether to render a
            blank form (`CREATE`) or load the existing testimonial via `GET /submissions/mine`
            (`EDIT`), per `email_lookup_hash` (decision 6).
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  mode: { type: string, enum: [CREATE, EDIT] }
                  testimonialId: { type: integer, nullable: true }
        "401":
          description: Wrong or expired code; attempts counter decremented (decision 17, NFR-VISITOR-OTP-BRUTEFORCE).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /submissions/session:
    get:
      tags: [submission]
      summary: Session ping for the visitor's login-protected pages (decision 23)
      security: [{ visitorSession: [] }]
      description: >
        Answers whether the visitor session is still alive; the request itself extends it.
        Called by the pages' session check (static/js/session-check.js) whenever their tab
        becomes visible again — on 401 or 403 the page reloads and so goes through the login
        redirect. Responses are not cacheable (`Cache-Control` includes `no-store`).
      responses:
        "204": { description: The visitor session is alive. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /submissions:
    post:
      tags: [submission]
      summary: Create a new testimonial (UC-CREATE-TESTIMONIAL)
      security: [{ visitorSession: [] }]
      description: >
        Topics are database-driven (decision 11), so the request cannot have one named field
        per topic. `payload` is a JSON part; photo files are separate multipart parts,
        cross-referenced from `payload.sections[].photos[].fileRef` by index. The exact
        multipart framing is an implementation detail; this schema fixes the logical shape
        only.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/TestimonialSubmission" }
      responses:
        "201":
          description: Saved with status PENDING.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  status: { type: string, enum: [PENDING] }
        "400":
          description: >
            Validation failure — no section filled in, missing recommendation score, a
            mandatory identity field missing, no country ("Please select your country."), a
            roll number not in the `AA00A000` shape or an admission year outside
            1959..current year (decision 10), data-processing consent unchecked (decision 7),
            photos on a section with a blank answer, a contact value not matching its type's
            `valuePattern` (decision 5), more photos on one section than the per-topic limit
            ("At most 5 photos per topic.") or on the testimonial than the total limit ("Too
            many photos: maximum is 50."; defaults, decision 22), more than 10 custom tags on
            one photo, or a multipart request the server can't parse (e.g. more parts than the
            configured part-count limit). Every one of these rules the request breaks is
            reported at once, as `field: message` pairs joined into `message`. Only once they
            all pass are the photo files themselves checked and converted, one at a time, so
            the first photo that fails is reported alone: over the size limit ("Photo exceeds
            the maximum allowed file size (20 MB)."), not a readable image ("Unsupported image
            format. Please upload JPEG, PNG, WebP, GIF, TIFF or BMP." — HEIC/HEIF and AVIF
            included), or above the pixel limit ("The image is too large: at most 250
            megapixels are allowed.").
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: A testimonial already exists for this visitor — use PUT /submissions/mine instead.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "413":
          description: >
            The multipart request exceeds the server's upload limits (per-part or per-request
            size; defaults 25MB and 1010MB, above the app's own 20 MB photo limit) before the
            application's own validation can run.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /admin/auth/otp/request:
    post:
      tags: [adminauth]
      summary: Request an admin login OTP (UC-ADMIN-OTP-REQUEST)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202":
          description: >
            Accepted regardless of whether the email matched, to avoid revealing which part
            of login failed (decision 4).
        "429":
          description: Too many OTP requests in a short window.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/auth/otp/verify:
    post:
      tags: [adminauth]
      summary: Verify an admin login OTP and establish a session (UC-ADMIN-OTP-VERIFY)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email: { type: string, format: email }
                code: { type: string, minLength: 6, maxLength: 6 }
      responses:
        "200":
          description: OTP correct — session cookie set.
          headers:
            Set-Cookie:
              schema: { type: string }
        "401":
          description: Wrong or expired code; attempts counter decremented (decision 4).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Established via /admin/auth/otp/verify; principal holds the admin role.
    visitorSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >
        Established via /submissions/otp/verify; principal holds the visitor role, scoped to
        the one email that was verified (decision 17).
...
  responses:
...
    Unauthorized:
      description: No valid session for the required role.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: A valid session, but of the other role (e.g. an admin session on a visitor endpoint).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    TestimonialSubmission:
      description: >
        Logical shape of the multipart submission body for POST /submissions and
        PUT /submissions/mine, and the response body of GET /submissions/mine.
      type: object
      required:
        [firstName, lastName, rollNumber, admissionYear, countryCode, recommendationScore,
         sections, dataProcessingConsent]
      properties:
        firstName: { type: string }
        lastName: { type: string }
        rollNumber:
          type: string
          pattern: "^[A-Za-z]{2}[0-9]{2}[A-Za-z][0-9]{3}$"
          description: >
            IITM roll number in the `AA00A000` shape, e.g. `CS21B001` (decision 10). Trimmed and
            uppercased before validation, so lowercase letters are accepted and stored uppercased.
        admissionYear:
          type: integer
          minimum: 1959
          description: Between 1959 and the current year inclusive (decision 10).
        countryCode: { type: string, minLength: 2, maxLength: 2 }
        recommendationScore: { type: integer, minimum: 0, maximum: 10 }
        sections:
          type: array
          minItems: 1
          description: >
            At least one filled section required (decision 12). Each references a live Topic by
            slug. A section whose `answer` is blank doesn't count as filled and is dropped, but
            only if it carries no photos — photos on a blank section are rejected with 400.
          items:
            type: object
            required: [topicSlug, answer]
            properties:
              topicSlug: { type: string }
              answer: { type: string }
              photos:
                type: array
                description: >
                  Indexes into the multipart file parts carrying this section's photos. At most
                  the configured per-topic limit per section and the total limit per testimonial,
                  kept and new photos counted together (defaults 5 and 50, decision 22). Each new
                  file is converted to WebP on save; the uploaded file itself is never stored.
                items:
                  type: object
                  properties:
                    fileRef:
                      type: string
                      description: >
                        On create, always names a multipart file part (a new upload). On edit
                        (PUT /submissions/mine), it is dual-purpose: if it matches an existing
                        photo's own `url` (as returned by GET /submissions/mine, in this same
                        section) that photo is kept unchanged except for `tags`; otherwise it
                        must name a multipart file part, i.e. a genuinely new upload. Any of the
                        section's existing photos whose `url` isn't referenced this way are
                        removed (decision 18).
                    tags:
                      type: array
                      items: { type: string }
                      maxItems: 10
        achievementSlugs:
          type: array
          items: { type: string }
          description: Achievement checkboxes ticked, from the current active Achievement list.
        contactMethods:
          type: array
          description: 0-N contact methods (decision 5), each independently marked public/private.
          items:
            type: object
            required: [typeSlug, value, isPublic]
            properties:
              typeSlug: { type: string, description: "ContactType slug, e.g. \"telegram\"" }
              value:
                type: string
                description: >
                  Trimmed, then must fully match the type's `valuePattern` when it has one
                  (decision 5); otherwise only non-blank.
              isPublic: { type: boolean, default: false }
        dataProcessingConsent:
          type: boolean
          description: Must be true, per decision 7 — request is rejected otherwise.
...
    ErrorResponse:
      description: Uniform error shape returned by @RestControllerAdvice (per CLAUDE.md baseline).
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        path: { type: string }